Refactor build query code
